// src/app/models/housingLocation.ts
export interface HousingLocation {
    id: number;
    photo: string;
    name: string;
    state:string;
    city: string;
    type: string;
    area: string;
    status:string;
    availableUnits: number;
    wifi: boolean;
    laundry: boolean;
    price: string
    readmore: string;
    amenities?: any;
    video: string;
    slider?: any;

}

  